back-crawling/requirements.txt
uvicorn==0.29.0
fastapi~=0.111.0
openai==1.23.6
pip==24.0
python-dotenv==1.0.1
selenium==4.20.0

starlette~=0.37.2
requests~=2.31.0
beautifulsoup4~=4.12.3